phase-3-specialization/rag-conversational-ai-assistant/requirements-azure.txt
# Azure App Service Optimized Requirements for RAG Conversational AI Assistant
# Based on successful patterns from intelligent-document-analysis deployment

# Core Framework (minimal for Azure)
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0

# Database (Azure PostgreSQL compatible + SQLite fallback)
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
aiosqlite==0.20.0

# AI/ML Libraries (optimized versions)
openai==1.58.1
anthropic==0.37.1
langchain==0.3.7
langchain-community==0.3.5
langchain-openai==0.2.5
langchain-anthropic==0.2.4
chromadb==0.5.20
sentence-transformers==3.2.1
transformers==4.46.3
torch==2.5.1
huggingface-hub==0.26.2

# Document Processing (essential only)
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
python-magic==0.4.27
pytesseract==0.3.13
Pillow==11.0.0

# Web Interface and API
streamlit==1.40.2
plotly==5.24.1
pandas==2.2.3
numpy==2.1.3

# HTTP and API
httpx==0.28.1
requests==2.32.3
aiofiles==24.1.0

# Async support
asyncio==3.4.3
asyncpg==0.30.0

# Configuration and Environment
python-dotenv==1.0.1
pyyaml==6.0.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# Monitoring and Logging (Azure compatible)
structlog==24.4.0
prometheus-client==0.21.1

# Azure specific
azure-storage-blob==12.23.0
azure-identity==1.19.0
azure-keyvault-secrets==4.8.0

# Production server optimizations
gunicorn==23.0.0
uvloop==0.21.0
httptools==0.6.4

# Memory optimization
pympler==0.9
memory-profiler==0.61.0

# Performance monitoring
psutil==6.1.0

# Redis for caching (Azure Redis compatible)
redis==5.2.0
aioredis==2.0.1

# Text processing optimizations
nltk==3.9.1
spacy==3.8.2
textstat==0.7.3

# Additional utilities for production
click==8.1.7
python-dateutil==2.9.0.post0
typing-extensions==4.12.2
